# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistem Informasi Kepegawaian dan Status Kehadiran API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
...
  /attendance:
    get:
      operationId: getAttendanceRecords
      tags: [attendance]
      summary: Get attendance records
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [izin, tidak_hadir, cuti, dinas]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendanceRecord
      tags: [attendance]
      summary: Create attendance record (izin/absensi)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRecord"
      responses:
        "201":
          description: Attendance record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
...
  /attendance/{id}:
    put:
      operationId: updateAttendanceRecord
      tags: [attendance]
      summary: Update attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRecord"
      responses:
        "200":
          description: Attendance record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
    delete:
      operationId: deleteAttendanceRecord
      tags: [attendance]
      summary: Delete attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Record deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /complaints:
    get:
      operationId: getComplaints
      tags: [complaints]
      summary: Get all complaints
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [baru, diproses, selesai, ditolak]
        - name: prioritas
          in: query
          schema:
            type: string
            enum: [rendah, sedang, tinggi]
        - name: kategori
          in: query
          schema:
            type: string
            enum: [produk, layanan, pengiriman, lainnya]
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Create new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaint"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    put:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaint"
      responses:
        "200":
          description: Complaint updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    delete:
      operationId: deleteComplaint
      tags: [complaints]
      summary: Delete complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    StatusPekerja:
      type: string
      enum: [Organik, TAD, TKJP, Security, Mitra Kerja, Driver, CS, Gardener]
    StatusBekerja:
      type: string
      enum: [Aktif, Mutasi, Pensiun, PHK, Mengundurkan Diri]
    Employee:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        nopek:
          type: string
        foto:
          type: string
          nullable: true
        jabatan:
          type: string
          nullable: true
        departemen:
          type: string
          nullable: true
        fungsi:
          type: string
          nullable: true
        lokasi_kerja:
          type: string
          nullable: true
        nomor_telepon:
          type: string
          nullable: true
        status_pekerja:
          $ref: "#/components/schemas/StatusPekerja"
          nullable: true
        status_bekerja:
          $ref: "#/components/schemas/StatusBekerja"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - nama
        - nopek
        - createdAt
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employee:
          $ref: "#/components/schemas/Employee"
        status:
          type: string
          enum: [izin, tidak_hadir, cuti, dinas]
        tanggal:
          type: string
        alasan:
          type: string
          nullable: true
        keterangan:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - status
        - tanggal
        - createdAt
...
    CreateAttendanceRecord:
      type: object
      properties:
        employeeId:
          type: integer
        status:
          type: string
          enum: [izin, tidak_hadir, cuti, dinas]
        tanggal:
          type: string
        alasan:
          type: string
          nullable: true
        keterangan:
          type: string
          nullable: true
      required:
        - employeeId
        - status
        - tanggal
...
    Complaint:
      type: object
      properties:
        id:
          type: integer
        namaPelanggan:
          type: string
        kontakPelanggan:
          type: string
          nullable: true
        kategori:
          type: string
          enum: [produk, layanan, pengiriman, lainnya]
        judul:
          type: string
        deskripsi:
          type: string
        status:
          type: string
          enum: [baru, diproses, selesai, ditolak]
        prioritas:
          type: string
          enum: [rendah, sedang, tinggi]
        tanggal:
          type: string
        penangananOleh:
          type: string
          nullable: true
        catatanPenanganan:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - namaPelanggan
        - kategori
        - judul
        - deskripsi
        - status
        - prioritas
        - tanggal
        - createdAt
    CreateComplaint:
      type: object
      properties:
        namaPelanggan:
          type: string
        kontakPelanggan:
          type: string
          nullable: true
        kategori:
          type: string
          enum: [produk, layanan, pengiriman, lainnya]
        judul:
          type: string
        deskripsi:
          type: string
        status:
          type: string
          enum: [baru, diproses, selesai, ditolak]
        prioritas:
          type: string
          enum: [rendah, sedang, tinggi]
        tanggal:
          type: string
        penangananOleh:
          type: string
          nullable: true
        catatanPenanganan:
          type: string
          nullable: true
      required:
        - namaPelanggan
        - kategori
        - judul
        - deskripsi
        - status
        - prioritas
        - tanggal
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        roleId:
          type: integer
        roleName:
          type: string
          nullable: true
        roleDescription:
          type: string
          nullable: true
        employeeId:
          type: integer
          nullable: true
        permissions:
          type: array
          items:
            type: string
      required:
        - id
        - username
        - roleId
        - permissions
...