# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Sistem Informasi Kepegawaian dan Status Kehadiran API
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintKategori } from "./complaintKategori";
import type { ComplaintPrioritas } from "./complaintPrioritas";
import type { ComplaintStatus } from "./complaintStatus";

export interface Complaint {
  id: number;
  namaPelanggan: string;
  kontakPelanggan?: string | null;
  kategori: ComplaintKategori;
  judul: string;
  deskripsi: string;
  status: ComplaintStatus;
  prioritas: ComplaintPrioritas;
  tanggal: string;
  penangananOleh?: string | null;
  catatanPenanganan?: string | null;
  createdAt: string;
}