# lib/api-zod/src/generated/types/createAttendanceRecord.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Sistem Informasi Kepegawaian dan Status Kehadiran API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAttendanceRecordStatus } from "./createAttendanceRecordStatus";

export interface CreateAttendanceRecord {
  employeeId: number;
  status: CreateAttendanceRecordStatus;
  tanggal: string;
  alasan?: string | null;
  keterangan?: string | null;
}