# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Sistem Informasi Kepegawaian dan Status Kehadiran API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  username: string;
  password: string;
}